Rentify 8 jwt login and registration (#13)
* Added JwtAuthFilter,JwtService and dependecy for Jwt in pom file

* Added JWT

* Updated JWT register

* Update authentication

* Updatte Login

* register page is added

confirm password added

* fix address validation in register page

* address removed from register

* Update register

* register form is working

phone is still not working

* Update Constrains in Register page

* Update validations in Register and Login

* auth route added

* Updated the Login to redirect to home and organized the classes into packages.

* login and register container size fixed

* login-register css fixed

 routes fixed

* Bug fix

* Resolve conflict

---------

+++ back-end/src/main/java/com/devminds/rentify/service/AddressService.java
@@ -22,7 +22,7 @@ public class AddressService {
         return addressRepository.findAll();
     }
 
-    public Address getAddressById(long id) {
+    public Address getAddressById(int id) {
         return addressRepository.findById(id)
                 .orElseThrow(() -> new AddressNotFoundException(String.format(ADDRESS_NOT_FOUND_MESSAGE, id)));
     }
@@ -39,7 +39,7 @@ public class AddressService {
         return addressRepository.save(address);
     }
 
-    public void deleteAddressById(long id) {
+    public void deleteAddressById(int id) {
         addressRepository.deleteById(id);
     }
 }
